+++ src/arcsf/data/README.md
@@ -1,10 +1,10 @@
 # Utils
-## Data Utils 
+## Data Utils
 ### `load_tofu`
  `load_tofu` is used to load the TOFU datasets compatible with the original source code, but with varying forget/retain sizes, and varying granularity (pending) of unlearning content.
 
  It outputs two datasets of the format/type:
-    
+
     Dataset({
         features: ['question', 'answer'],
         num_rows: N # N depends on the chosen forget fraction
@@ -12,7 +12,7 @@
 
 The first output is the forget set, and the second is the retain set.
 
-It can (hopefully) be integrated into the source code in the following way, using the Text: 
+It can (hopefully) be integrated into the source code in the following way, using the Text:
 
     class TextForgetDatasetQA(Dataset):
         def __init__(self, data_path, tokenizer, model_family,  max_length=512, split = "forget10", loss_type="idk"):
@@ -27,4 +27,4 @@ It can (hopefully) be integrated into the source code in the following way, usin
 
             ...
 
-The `self.forget_data` and `self.retain_data` attributes should be replaced with those defined using `load_tofu`. There is no need to specify the `['train']` split, since the `load_tofu` function does not create a datasetDict (though there is scope to change this if necessary for improved integrability). 
+The `self.forget_data` and `self.retain_data` attributes should be replaced with those defined using `load_tofu`. There is no need to specify the `['train']` split, since the `load_tofu` function does not create a datasetDict (though there is scope to change this if necessary for improved integrability).
